backend/requirements.txt
fastapi
uvicorn
azure-search-documents
openai
python-dotenv
